pyproject.toml
[tool.poetry]
name = "tentacruel"
version = "0.1.0"
description = "Software library and GUI for Denon HEOS speakers and amplifiers"
authors = ["Paul Houle <paul.houle@ontology2.com>"]

[tool.poetry.dev-dependencies]
pytest = "^3.0"
[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^3.13"
skyfield = "^1.10"
python-dateutil = "^2.8"
phue = "^1.1"
boto3 = "^1.9"
python-arango = "^4.4"
pytz = "^2018.9"
aio-pika = "^5.3"
mutagen = "^1.42"
multidict = "^4.5"
imageio = "^2.5"
bs4 = "^0.0.1"
lxml = "^4.3"
jinja2 = "^2.10"
metar = "^1.7"
aiohttp = "^3.5"
zstd = "^1.4"
nanoleafapi = "^1.1.0"
frozendict = "^1.2"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"